Move PyLops v2 from dev to master

+++ pytests/test_transpose.py
@@ -1,13 +1,12 @@
 import numpy as np
 import pytest
 from numpy.testing import assert_equal
-from scipy.sparse.linalg import lsqr
 
 from pylops.basicoperators import Transpose
 from pylops.utils import dottest
 
-par1 = {"ny": 21, "nx": 11, "nt": 20, "imag": 0, "dtype": "float32"}  # real
-par2 = {"ny": 21, "nx": 11, "nt": 20, "imag": 1j, "dtype": "complex64"}  # complex
+par1 = {"ny": 21, "nx": 11, "nt": 20, "imag": 0, "dtype": "float64"}  # real
+par2 = {"ny": 21, "nx": 11, "nt": 20, "imag": 1j, "dtype": "complex128"}  # complex
 
 np.random.seed(10)
